+++ libbsdfetch/http.c
@@ -454,7 +454,7 @@ http_cmd(conn_t *conn, const char *fmt, ...)
 		return (-1);
 	}
 
-	ED_TRACE(("cmd: <%.*s> (%d)\n", len-2, msg, len))
+	ED_TRACE(("cmd: <%.*s> (%d)\n", (int)(len - 2), msg, len))
 	r = fetch_write(conn, msg, len);
 	free(msg);
